Fix venv entering
Fixes a bug where if

1. you have invoke installed with pyenv, and
2. run invoke while in a venv without invoke

then pyenv will add its paths to the head of PATH, invoke will not enter
the venv, and then complain because the Python in PATH is not the same
as in the venv.

Now invoke will always put the venv to the top of the PATH, so if a path
with Python has been added to PATH since entering the venv the venv
Python will still be preferred.