Change virtualenv task to set PATH to venv only
We found a bug where if pip-sync wasn't installed in the venv but was
installed in a wider environment, then pip-sync would run in that
environment and trash your python installation. This is not desirable.

This commit changes the task `virtualenv` so that now the venv bin
folder will be the only directory in `PATH`; the downside of this is that
the only commands usable for tasks in this repo will be Python commands,
the upside is that we won't accidentally trash people's installations.