Fix bootstrap of digitalmarketplace-developer-tools
This package may from time to time add dependencies to itself and use
those dependencies in its task file; installing developer tools from
PyPI like we would for another repo doesn't work in that case because
the dependencies won't be installed.

Instead we install digitalmarketplace-developer-tools from the repo, so
we're always using the latest version.